requirements.txt
# Core Framework Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
pydantic>=2.5.0
python-multipart>=0.0.6

# AI and LLM Dependencies
langchain>=0.1.0
langchain-community>=0.0.38
langchain-core>=0.1.0
openai>=1.3.0
groq>=0.4.0
transformers>=4.36.0
torch>=2.0.0
sentence-transformers>=2.2.0
spacy>=3.4.0

# Vector Storage and Search
faiss-cpu>=1.7.0
chromadb>=0.4.0

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.13.0

# Data Processing and Analysis
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Report Generation
reportlab>=4.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
jinja2>=3.1.0

# Utilities and Helpers
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
aiofiles>=23.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Logging and Monitoring
loguru>=0.7.0
prometheus-client>=0.19.0

# File Processing
openpyxl>=3.1.0

# Date and Time
python-dateutil>=2.8.0

# HTTP Client
httpx>=0.25.0

# Environment and Configuration
click>=8.1.0
rich>=13.7.0
